+++ projects/rocprofiler-compute/src/argparser.py
@@ -24,6 +24,7 @@
 ##############################################################################
 
 import argparse
+import os
 import re
 from pathlib import Path
 from typing import Optional
@@ -351,7 +352,9 @@ Examples:
         type=str,
         dest="rocprofiler_sdk_library_path",
         required=False,
-        default="/opt/rocm/lib/librocprofiler-sdk.so",
+        default=str(
+            Path(os.getenv("ROCM_PATH", "/opt/rocm")) / "lib/librocprofiler-sdk.so"
+        ),
         help="\t\t\tSet the path to rocprofiler SDK library.",
     )
     profile_group.add_argument(