+++ projects/rocprofiler-compute/src/rocprof_compute_soc/soc_gfx942.py
@@ -37,8 +37,6 @@ class gfx942_soc(OmniSoC_Base):
         super().__init__(args, mspec)
         self.set_arch("gfx942")
         self.set_compatible_profilers([
-            "rocprofv1",
-            "rocprofv2",
             "rocprofv3",
             "rocprofiler-sdk",
         ])